combined all exomol files
